Move collect_upvars into its own file.
